[PP-13] Creating REST API for Account Value Calculation
Currently has two versions of the backend code, one with sync calls and one with async. Will remove the sync code when we verify the async stuff works properly.
Added Copilot-generated tests (which all pass), and verfied through Postman that the endpoint returns a real value.

Handles PP-32 with the async code